requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Data processing
kaggle>=1.5.16
python-dotenv>=0.19.0
pathlib2>=2.3.0

# PDF generation
reportlab>=3.6.0

# Hugging Face Transformers for semantic search and text generation
transformers>=4.21.0
torch>=1.12.0
accelerate>=0.20.0
sentence-transformers>=2.2.0

# Semantic search and embeddings
scikit-learn>=1.1.0
scipy>=1.9.0

# Async support
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0

# Logging and monitoring
prometheus-client>=0.14.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 